extracts Branding to its own folder
