+++ components/src/Branding/index.js
@@ -1,0 +1,1 @@
+export { default as Branding } from "./Branding";
